+++ src/test/groovy/ru/vyarus/java/generics/resolver/ContextTypesResolutionTest.groovy
@@ -1,7 +1,6 @@
 package ru.vyarus.java.generics.resolver
 
 import ru.vyarus.java.generics.resolver.context.GenericsContext
-import ru.vyarus.java.generics.resolver.context.container.GenericArrayTypeImpl
 import ru.vyarus.java.generics.resolver.context.container.ParameterizedTypeImpl
 import ru.vyarus.java.generics.resolver.support.Lvl2Base1
 import ru.vyarus.java.generics.resolver.support.Model
@@ -21,7 +20,7 @@ class ContextTypesResolutionTest extends Specification {
         println(context)
 
         expect: "resolving type"
-        context.resolveType(Lvl2Base1.getMethod("doSomth5").getGenericReturnType()) == new GenericArrayTypeImpl(Model)
+        context.resolveType(Lvl2Base1.getMethod("doSomth5").getGenericReturnType()) == Model[]
         context.resolveType(Lvl2Base1.getMethod("doSomth6").getGenericReturnType()) == new ParameterizedTypeImpl(Map, Model, Model)
     }