+++ src/test/groovy/ru/vyarus/java/generics/resolver/support/array/ArRoot.groovy
@@ -6,5 +6,5 @@ import ru.vyarus.java.generics.resolver.support.Model
  * @author Vyacheslav Rusakov 
  * @since 15.12.2014
  */
-interface ArRoot extends ArBase<Model, ? extends Model, ? super Model> {
+interface ArRoot extends ArBase<List<Model>, ? extends Model, ? super Model> {
 }